+++ packages/eslint-config-vtex-react/rules/react-hooks.js
@@ -1,4 +1,6 @@
 module.exports = {
+  extends: ['plugin:react-hooks/recommended'],
+  plugins: ['react-hooks'],
   rules: {
     // Enforce Rules of Hooks
     // https://github.com/facebook/react/blob/c11015ff4f610ac2924d1fc6d569a17657a404fd/packages/eslint-plugin-react-hooks/src/RulesOfHooks.js